moved semantics inside engine
