+++ bitloops_cli/src/commands/devql/support.rs
@@ -1,4 +1,3 @@
 include!("ingestion.rs");
-include!("semantic_features.rs");
 include!("query_pipeline.rs");
 include!("db_utils.rs");
