+++ bitloops_cli/src/engine/semantic_features/mod.rs
@@ -1,0 +1,11 @@
+pub mod semantic_features;
+
+pub use semantic_features::{
+    NoopSemanticSummaryProvider, PreStageArtefactRow, SemanticFeatureIndexState,
+    SemanticFeatureInput, SemanticSummaryCandidate, SemanticSummaryProvider,
+    SemanticSummaryProviderConfig, SemanticSummarySource,
+    build_semantic_feature_inputs_from_artefacts, build_semantic_feature_rows,
+    build_semantic_summary_provider, load_pre_stage_artefacts_for_blob,
+    resolve_semantic_summary_endpoint, semantic_features_require_reindex,
+    upsert_semantic_feature_rows,
+};
